Simplify container maintenance
Alpine provides up-to-date clamav packages and it is not necessary to
have involved locking mechanisms in the lauch scripts to handle the av
database first-fetch and update.

Rsyslog is included here because neither clamd nor freshclam can write
directly to `/dev/stdout`-both try to open the device in append mode,
which raises an error.